+++ interfaces/HO-Portal/src/app/models/person.model.ts
@@ -119,9 +119,10 @@ export const PA_STATUS_ORDER = [
   { id: 7, value: RegistrationStatus.included },
   { id: 8, value: RegistrationStatus.inclusionEnded },
   { id: 9, value: RegistrationStatus.rejected },
-  { id: 10, value: RegistrationStatus.noLongerEligible },
-  { id: 11, value: RegistrationStatus.registeredWhileNoLongerEligible },
-  { id: 12, value: RegistrationStatus.completed },
+  { id: 10, value: RegistrationStatus.paused },
+  { id: 11, value: RegistrationStatus.noLongerEligible },
+  { id: 12, value: RegistrationStatus.registeredWhileNoLongerEligible },
+  { id: 13, value: RegistrationStatus.completed },
 ];
 
 export enum PersonDefaultAttributes {