+++ interfaces/HO-Portal/src/app/pages/dashboard/dashboard.module.ts
@@ -3,6 +3,7 @@ import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { RouterModule, Routes } from '@angular/router';
 import { IonicModule } from '@ionic/angular';
+import { DashboardIframeComponent } from 'src/app/components/dashboard-iframe/dashboard-iframe.component';
 import { MetricsStatesComponent } from 'src/app/program/metrics-states/metrics-states.component';
 import { MetricsTotalsComponent } from 'src/app/program/metrics-totals/metrics-totals.component';
 import { SharedModule } from 'src/app/shared/shared.module';
@@ -22,6 +23,7 @@ const routes: Routes = [
     IonicModule,
     SharedModule,
     RouterModule.forChild(routes),
+    DashboardIframeComponent,
   ],
   declarations: [DashboardPage, MetricsStatesComponent, MetricsTotalsComponent],
 })