+++ interfaces/HO-Portal/src/app/services/programs-service-api.service.ts
@@ -27,7 +27,7 @@ import {
 } from '../models/program.model';
 import { RegistrationChangeLog } from '../models/registration-change-log.model';
 import { Transaction } from '../models/transaction.model';
-import { User } from '../models/user.model';
+import { Role, TableData, User } from '../models/user.model';
 import { ImportResult } from '../program/bulk-import/bulk-import.component';
 import { arrayToXlsx } from '../shared/array-to-xlsx';
 import { ApiService } from './api.service';
@@ -135,9 +135,10 @@ export class ProgramsServiceApiService {
     programId: number | string,
     phase: ProgramPhase,
   ): Promise<PaTableAttribute[]> {
+    const phaseString = phase ? phase : '';
     return this.apiService.get(
       environment.url_121_service_api,
-      `/programs/${programId}/pa-table-attributes/${phase}`,
+      `/programs/${programId}/pa-table-attributes/${phaseString}`,
     );
   }
 
@@ -489,7 +490,7 @@ export class ProgramsServiceApiService {
   ): Promise<PhysicalCard[]> {
     const response = await this.apiService.get(
       environment.url_121_service_api,
-      `/programs/${programId}/fsp-integration/intersolve-visa/wallets?referenceId=${referenceId}`,
+      `/programs/${programId}/financial-service-providers/intersolve-visa/wallets?referenceId=${referenceId}`,
     );
 
     return !!response && !!response.wallets ? response.wallets : [];
@@ -501,7 +502,7 @@ export class ProgramsServiceApiService {
   ): Promise<any> {
     const res = await this.apiService.put(
       environment.url_121_service_api,
-      `/programs/${programId}/fsp-integration/intersolve-visa/customers/${referenceId}/wallets`,
+      `/programs/${programId}/financial-service-providers/intersolve-visa/customers/${referenceId}/wallets`,
       {},
     );
 
@@ -515,7 +516,7 @@ export class ProgramsServiceApiService {
   ): Promise<any> {
     return await this.apiService.post(
       environment.url_121_service_api,
-      `/programs/${programId}/fsp-integration/intersolve-visa/wallets/${tokenCode}/${
+      `/programs/${programId}/financial-service-providers/intersolve-visa/wallets/${tokenCode}/${
         block ? 'block' : 'unblock'
       }`,
       {},
@@ -628,6 +629,14 @@ export class ProgramsServiceApiService {
     message: string,
   ): Promise<any> {
     return this.updatePaStatus('reject', programId, referenceIds, message);
+  }
+
+  pause(
+    programId: number | string,
+    referenceIds: string[],
+    message: string,
+  ): Promise<any> {
+    return this.updatePaStatus('pause', programId, referenceIds, message);
   }
 
   sendMessage(
@@ -795,5 +804,13 @@ export class ProgramsServiceApiService {
       `/programs/${programId}/metrics/registration-status`,
       false,
     );
+  }
+
+  getAllUsers(): Promise<TableData[] | null> {
+    return this.apiService.get(environment.url_121_service_api, '/users');
+  }
+
+  getRoles(): Promise<Role[] | null> {
+    return this.apiService.get(environment.url_121_service_api, '/roles');
   }
 }