+++ services/121-service/src/programs/programs.service.ts
@@ -75,6 +75,15 @@ export class ProgramService {
       );
     }
 
+    let includeMetricsUrl = false;
+    if (userId) {
+      includeMetricsUrl = await this.userService.canActivate(
+        [PermissionEnum.ProgramMetricsREAD],
+        programId,
+        userId,
+      );
+    }
+
     let relations = [
       'programQuestions',
       'financialServiceProviders',
@@ -104,6 +113,11 @@ export class ProgramService {
 
       // TODO: Get these attributes from some enum or something
       program['filterableAttributes'] = this.getFilterableAttributes(program);
+
+      if (!includeMetricsUrl) {
+        delete program.monitoringDashboardUrl;
+        delete program.evaluationDashboardUrl;
+      }
     }
     // TODO: REFACTOR: use DTO to define (stable) structure of data to return (not sure if transformation should be done here or in controller)
     return program;
@@ -161,14 +175,15 @@ export class ProgramService {
 
   public async getCreateProgramDto(
     programId: number,
+    userId: number,
   ): Promise<CreateProgramDto> {
-    const programEntity = await this.findOne(programId);
+    const programEntity = await this.findOne(programId, userId);
     if (!programEntity) {
       const errors = `No program found with id ${programId}`;
       throw new HttpException({ errors }, HttpStatus.NOT_FOUND);
     }
 
-    return {
+    const programDto: CreateProgramDto = {
       published: programEntity.published,
       validation: programEntity.validation,
       phase: programEntity.phase,
@@ -233,6 +248,13 @@ export class ProgramService {
       languages: programEntity.languages,
       enableMaxPayments: programEntity.enableMaxPayments,
     };
+    if (programEntity.monitoringDashboardUrl) {
+      programDto.monitoringDashboardUrl = programEntity.monitoringDashboardUrl;
+    }
+    if (programEntity.evaluationDashboardUrl) {
+      programDto.evaluationDashboardUrl = programEntity.evaluationDashboardUrl;
+    }
+    return programDto;
   }
 
   public async getPublishedPrograms(): Promise<ProgramsRO> {
@@ -360,6 +382,8 @@ export class ProgramService {
     program.fullnameNamingConvention = programData.fullnameNamingConvention;
     program.languages = programData.languages;
     program.enableMaxPayments = programData.enableMaxPayments;
+    program.monitoringDashboardUrl = programData.monitoringDashboardUrl;
+    program.evaluationDashboardUrl = programData.evaluationDashboardUrl;
 
     const queryRunner = this.dataSource.createQueryRunner();
     await queryRunner.startTransaction();