+++ CHANGELOG.md
@@ -5,22 +5,57 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/global-121/121-platform/compare/v1.97.2...master)
+## [Unreleased](https://github.com/global-121/121-platform/compare/v1.100.0...master)
 
 ---
+## [1.100.0](https://github.com/global-121/121-platform/compare/v1.99.1...v1.100.0) - 2023-09-21
+
+### Added
+- Allow external integrations to update the reference Id using patch
+- Admins can see an overview of all users
+- Admins can see an overview of all roles
+- The possiblity to add an url to a powerbi dashboard that can be shown in the evaluation page and in the dashboard page
+- Removed old code for the espocrm integration
+
+
+## [1.99.1](https://github.com/global-121/121-platform/compare/v1.99.0...v1.99.1) - 2023-09-19
+
+### Fixed
+
+- Reuse valid Intersolve Visa token instead of constantly creating new one
+
+## [1.99.0](https://github.com/global-121/121-platform/compare/v1.98.1...v1.99.0) - 2023-09-15
+
+### Added
+
+- Registration status paused
+- See PA number in profile page
+
+## [1.98.1](https://github.com/global-121/121-platform/compare/v1.98.0...v1.98.1) - 2023-09-13
+
+### Fixed
+
+- Get pa attributes without phase fix
+
+## [1.98.0](https://github.com/global-121/121-platform/compare/v1.97.2...v1.98.0) - 2023-09-13
+
+### Added
+
+- Let recipient page in portal use the profile page
+- Api endpoint to post program questions and custom attributes
+- Sync with intersolve for registrations that no longer have the Visa FSP but still have a Visa card
+
 ## [1.97.2](https://github.com/global-121/121-platform/compare/v1.97.1...v1.97.2) - 2023-09-12
 
 ### Fixed
 
 - Merging 2 registrations with registration changelog
 
-
 ## [1.97.1](https://github.com/global-121/121-platform/compare/v1.97.0...v1.97.1) - 2023-09-11
 
 ### Fixed
 
 - Fixed reissue inactive blocked wallet
-
 
 ## [1.97.0](https://github.com/global-121/121-platform/compare/v1.97.0...v1.96.0) - 2023-08-24
 
@@ -33,10 +68,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Show payment "status" without payment create permission
 
 ### Changed
+
 - Nestjs upgrade to v10
 - Use 'reserved' transactions in lastUsedDate Visa debit card
 
 ### Fixed
+
 - Export list of unsused voucher button
 
 ## [1.96.0](https://github.com/global-121/121-platform/compare/v1.95.1...v1.96.0) - 2023-08-16